Summary: Functions used to run selective equalization based on the SQNR obtained from running the Numeric Suite. After running the Numeric Suite between the equalized and float model, we will get the SQNR between the two models and construct an equalization_qconfig_dict that specifies to only equalize the layers with the highest quantization errors. Then, we will run equalization on only those layers.

How to run:
```
selectively_equalized_model = selctive_equalization_with_numeric_suite(float_model, qconfig_dict, input, num_layers_to_equalize)
```

Test Plan: `python test/test_quantization.py TestEqualizeFx.test_selective_equalization`

Reviewers:

Subscribers:

Tasks:

Tags:

Differential Revision: [D29796950](https://our.internmc.facebook.com/intern/diff/D29796950)

[ghstack-poisoned]

+++ aten/src/ATen/native/SpectralOps.cpp
@@ -208,10 +208,10 @@ ShapeAndDims canonicalize_fft_shape_and_dim_args(
 
   if (shape) {
     // Has shape, may have dim
-    TORCH_CHECK(!dim || static_cast<int64_t>(dim->size()) == shape->size(),
+    TORCH_CHECK(!dim ||
+                dim->size() == shape->size(),
                 "When given, dim and shape arguments must have the same length");
-    // NOLINTNEXTLINE(clang-diagnostic-sign-compare)
-    TORCH_CHECK(shape->size() <= input_dim,
+    TORCH_CHECK(static_cast<int64_t>(shape->size()) <= input_dim,
                 "Got shape with ", shape->size(), " values but input tensor "
                 "only has ", input_dim, " dimensions.");
     const int64_t transform_ndim = shape->size();
@@ -1048,7 +1048,6 @@ void _fft_fill_with_conjugate_symmetry_(const Tensor& input, IntArrayRef dim_) {
       mirror_dims, signal_half_sizes, in_strides, in_data, out_strides, out_data);
 }
 
-// NOLINTNEXTLINE(cppcoreguidelines-avoid-non-const-global-variables)
 DEFINE_DISPATCH(fft_fill_with_conjugate_symmetry_stub);
 
 }} // at::native
